next step

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/CustomOp.java
@@ -37,7 +37,7 @@ public interface CustomOp {
 
     INDArray[] inputArguments();
 
-    int[] iArgs();
+    long[] iArgs();
 
     double[] tArgs();
 
@@ -48,7 +48,7 @@ public interface CustomOp {
 
     void removeIArgument(Integer arg);
 
-    Integer getIArgument(int index);
+    Long getIArgument(int index);
 
     int numIArguments();
